Moved Material test API V1 to API V1 internal namespace. Material test controller makes use of API V3 Material Representers.
