+++ server/webapp/WEB-INF/rails.new/spec/controllers/api_v1/admin/internal/material_test_controller_spec.rb
@@ -16,7 +16,7 @@
 
 require 'spec_helper'
 
-describe ApiV1::Admin::MaterialTestController do
+describe ApiV1::Admin::Internal::MaterialTestController do
   include ApiHeaderSetupTeardown, ApiV1::ApiVersionHelper
 
   describe :test do
@@ -148,7 +148,7 @@ describe ApiV1::Admin::MaterialTestController do
       describe :with_header do
 
         it 'should route to test action of the material_test controller' do
-          expect(:post => 'api/admin/material_test').to route_to(action: 'test', controller: 'api_v1/admin/material_test')
+          expect(:post => 'api/admin/internal/material_test').to route_to(action: 'test', controller: 'api_v1/admin/internal/material_test')
         end
       end
       describe :without_header do
@@ -156,8 +156,8 @@ describe ApiV1::Admin::MaterialTestController do
           teardown_header
         end
         it 'should not route to test action of material_test controller without header' do
-          expect(:post => 'api/admin/material_test').to_not route_to(action: 'test', controller: 'api_v1/admin/material_test')
-          expect(:post => 'api/admin/material_test').to route_to(controller: 'application', action: 'unresolved', url: 'api/admin/material_test')
+          expect(:post => 'api/admin/internal/material_test').to_not route_to(action: 'test', controller: 'api_v1/admin/internal/material_test')
+          expect(:post => 'api/admin/internal/material_test').to route_to(controller: 'application', action: 'unresolved', url: 'api/admin/internal/material_test')
         end
       end
     end
