+++ server/webapp/WEB-INF/rails.new/spec/new_javascripts/models/pipeline_configs/materials_spec.js
@@ -151,7 +151,7 @@ define([
       describe('post', function () {
         it('should post to material_test url', function () {
           expect(requestArgs.method).toBe('POST');
-          expect(requestArgs.url).toBe('/go/api/admin/material_test');
+          expect(requestArgs.url).toBe('/go/api/admin/internal/material_test');
         });
 
         it('should post required headers', function () {
